Scripts: Split `npm run package` from `npm run build`.
This makes it possible to only do a test build (without packaging for
Electron).  To get the previous behaviour, run `npm run build` followed by
`npm run package`.

This also ensures `npm run package -- --win=zip` no longer builds the
Windows installer (as it used to).
