+++ .github/workflows/upgrade-generate.yaml
@@ -40,9 +40,8 @@ jobs:
       if: matrix.arch == 'aarch64' && matrix.platform == 'mac'
       run: echo "M1=1" >> "${GITHUB_ENV}"
     - run: npm ci
-    - run: npm run build -- --${{ matrix.platform }} --publish=never
-      env:
-        RD_FEAT_WIX: '1' # Will be optional soon
+    - run: npm run build
+    - run: npm run package -- --${{ matrix.platform }} --publish=never
     - name: Upload Windows exe installer
       if: runner.os == 'Windows'
       uses: actions/upload-artifact@v3
